int -> int64_t

+++ pulp/0.2/init_nonrandom.cpp
@@ -45,23 +45,23 @@
 
 using namespace std;
 
-extern int seed;
+extern int64_t seed;
 
 
 // multi-source bfs with random start points
-int* init_nonrandom(pulp_graph_t& g, int num_parts, int* parts)
+int* init_nonrandom(pulp_graph_t& g, int64_t num_parts, int64_t* parts)
 {
-  int num_verts = g.n;
-  int* queue = new int[num_verts*QUEUE_MULTIPLIER];
-  int* queue_next = new int[num_verts*QUEUE_MULTIPLIER];
-  int queue_size = num_parts;
-  int next_size = 0;
+  int64_t num_verts = g.n;
+  int64_t* queue = new int64_t[num_verts*QUEUE_MULTIPLIER];
+  int64_t* queue_next = new int64_t[num_verts*QUEUE_MULTIPLIER];
+  int64_t queue_size = num_parts;
+  int64_t next_size = 0;
 
 #pragma omp parallel
 {  
-  int thread_queue[ THREAD_QUEUE_SIZE ];
-  int thread_queue_size = 0;
-  int thread_start;
+  int64_t thread_queue[ THREAD_QUEUE_SIZE ];
+  int64_t thread_queue_size = 0;
+  int64_t thread_start;
 
   xs1024star_t xs;
   xs1024star_seed((unsigned long)(seed + omp_get_thread_num()), &xs);
@@ -74,7 +74,7 @@ int* init_nonrandom(pulp_graph_t& g, int num_parts, int* parts)
 {
   for (int i = 0; i < num_parts; ++i)
   {
-    int vert = (int)xs1024star_next(&xs) % num_verts;
+    int64_t vert = (int)xs1024star_next(&xs) % num_verts;
     while (parts[vert] != -1) {vert = (int)xs1024star_next(&xs) % num_verts;}
     parts[vert] = i;
     queue[i] = vert;
@@ -86,13 +86,13 @@ int* init_nonrandom(pulp_graph_t& g, int num_parts, int* parts)
 #pragma omp for schedule(guided) nowait
     for (int i = 0; i < queue_size; ++i)
     {
-      int vert = queue[i];
-      int part = parts[vert];
+      int64_t vert = queue[i];
+      int64_t part = parts[vert];
       long out_degree = out_degree(g, vert);
-      int* outs = out_vertices(g, vert);
+      int64_t* outs = out_vertices(g, vert);
       for (long j = 0; j < out_degree; ++j)
       {
-        int out = outs[j];
+        int64_t out = outs[j];
         if (parts[out] == -1)
         {
           parts[out] = part;
@@ -123,7 +123,7 @@ int* init_nonrandom(pulp_graph_t& g, int num_parts, int* parts)
 
 #pragma omp single
 {
-    int* temp = queue;
+    int64_t* temp = queue;
     queue = queue_next;
     queue_next = temp;
 
@@ -153,22 +153,22 @@ int* init_nonrandom(pulp_graph_t& g, int num_parts, int* parts)
 
 // multi-source bfs with random start points
 // constrain maximal size of part, randomly assign any remaining
-int* init_nonrandom_constrained(pulp_graph_t& g, int num_parts, int* parts)
+int* init_nonrandom_constrained(pulp_graph_t& g, int64_t num_parts, int64_t* parts)
 {
-  int num_verts = g.n;
+  int64_t num_verts = g.n;
 
-  int* queue = new int[num_verts*QUEUE_MULTIPLIER];
-  int* queue_next = new int[num_verts*QUEUE_MULTIPLIER];
-  int* part_sizes = new int[num_parts];
-  int queue_size = num_parts;
-  int next_size = 0;
-  int max_part_size = num_verts / num_parts * 2;
+  int64_t* queue = new int64_t[num_verts*QUEUE_MULTIPLIER];
+  int64_t* queue_next = new int64_t[num_verts*QUEUE_MULTIPLIER];
+  int64_t* part_sizes = new int64_t[num_parts];
+  int64_t queue_size = num_parts;
+  int64_t next_size = 0;
+  int64_t max_part_size = num_verts / num_parts * 2;
 
 #pragma omp parallel
 {  
-  int thread_queue[ THREAD_QUEUE_SIZE ];
-  int thread_queue_size = 0;
-  int thread_start;
+  int64_t thread_queue[ THREAD_QUEUE_SIZE ];
+  int64_t thread_queue_size = 0;
+  int64_t thread_start;
 
   xs1024star_t xs;
   xs1024star_seed((unsigned long)(seed + omp_get_thread_num()), &xs);
@@ -181,7 +181,7 @@ int* init_nonrandom_constrained(pulp_graph_t& g, int num_parts, int* parts)
 {
   for (int i = 0; i < num_parts; ++i)
   {
-    int vert = (int)xs1024star_next(&xs) % num_verts;
+    int64_t vert = (int)xs1024star_next(&xs) % num_verts;
     while (parts[vert] != -1) {vert = xs1024star_next(&xs) % num_verts;}
     parts[vert] = i;
     queue[i] = vert;
@@ -194,13 +194,13 @@ int* init_nonrandom_constrained(pulp_graph_t& g, int num_parts, int* parts)
 #pragma omp for schedule(guided) nowait
     for (int i = 0; i < queue_size; ++i)
     {
-      int vert = queue[i];
-      int part = parts[vert];
+      int64_t vert = queue[i];
+      int64_t part = parts[vert];
       long out_degree = out_degree(g, vert);
-      int* outs = out_vertices(g, vert);
+      int64_t* outs = out_vertices(g, vert);
       for (long j = 0; j < out_degree; ++j)
       {
-        int out = outs[j];
+        int64_t out = outs[j];
         if (parts[out] == -1)
         {
           if (part_sizes[part] < max_part_size)
@@ -237,7 +237,7 @@ int* init_nonrandom_constrained(pulp_graph_t& g, int num_parts, int* parts)
 
 #pragma omp single
 {
-    int* temp = queue;
+    int64_t* temp = queue;
     queue = queue_next;
     queue_next = temp;
